catalog: longs are now being used in the database.

+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/db/mongodb/CatalogMongoCohortDBAdaptor.java
@@ -45,7 +45,7 @@ public class CatalogMongoCohortDBAdaptor extends CatalogMongoDBAdaptor implement
         dbAdaptorFactory.getCatalogStudyDBAdaptor().checkStudyId(studyId);
         checkCohortNameExists(studyId, cohort.getName());
 
-        int newId = dbAdaptorFactory.getCatalogMetaDBAdaptor().getNewAutoIncrementId();
+        long newId = dbAdaptorFactory.getCatalogMetaDBAdaptor().getNewAutoIncrementId();
         cohort.setId(newId);
 
         Document cohortObject = cohortConverter.convertToStorageType(cohort);
